Fix UI issues in security tab of chatflow configuration dialog
