+++ packages/ui/src/ui-component/extended/AllowedDomains.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { enqueueSnackbar as enqueueSnackbarAction, closeSnackbar as closeSnackbarAction, SET_CHATFLOW } from '@/store/actions'
 
 // material-ui
-import { Button, IconButton, OutlinedInput, Box, List, InputAdornment, Typography } from '@mui/material'
+import { Button, IconButton, OutlinedInput, Box, List, InputAdornment, Stack, Typography } from '@mui/material'
 import { IconX, IconTrash, IconPlus } from '@tabler/icons-react'
 
 // Project import
@@ -118,22 +118,15 @@ const AllowedDomains = ({ dialogProps }) => {
     }, [dialogProps])
 
     return (
-        <>
-            <Box>
-                <Box
-                    sx={{
-                        display: 'flex',
-                        flexDirection: 'column'
-                    }}
-                >
-                    <Typography sx={{ mb: 1 }}>
-                        Allowed Domains
-                        <TooltipWithParser
-                            style={{ mb: 1, mt: 2, marginLeft: 10 }}
-                            title={'Your chatbot will only work when used from the following domains.'}
-                        />
-                    </Typography>
-                </Box>
+        <Stack direction='column' spacing={2} sx={{ alignItems: 'start' }}>
+            <Typography variant='h4'>
+                Allowed Domains
+                <TooltipWithParser
+                    style={{ mb: 1, mt: 2, marginLeft: 10 }}
+                    title={'Your chatbot will only work when used from the following domains.'}
+                />
+            </Typography>
+            <Stack direction='column' spacing={2} sx={{ width: '100%' }}>
                 <List>
                     {inputFields.map((origin, index) => {
                         return (
@@ -177,10 +170,8 @@ const AllowedDomains = ({ dialogProps }) => {
                         )
                     })}
                 </List>
-            </Box>
-            <Box sx={{ pt: 2, pb: 2 }}>
-                <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'flex-start' }}>
-                    <Typography sx={{ mb: 1 }}>
+                <Stack direction='column' spacing={1}>
+                    <Typography>
                         Error Message
                         <TooltipWithParser
                             style={{ mb: 1, mt: 2, marginLeft: 10 }}
@@ -198,12 +189,12 @@ const AllowedDomains = ({ dialogProps }) => {
                             setErrorMessage(e.target.value)
                         }}
                     />
-                </div>
-            </Box>
+                </Stack>
+            </Stack>
             <StyledButton variant='contained' onClick={onSave}>
                 Save
             </StyledButton>
-        </>
+        </Stack>
     )
 }
